requirements.txt
fastapi==0.68.1
starlette==0.14.2
pydantic==1.8.2
typing-extensions==3.10.0.2
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
SQLAlchemy==1.4.23
psycopg2_binary
SQLAlchemy_utils==0.37.8
requests~=2.31.0
python-dotenv~=1.0.0
werkzeug~=3.0.1
phonenumbers~=8.13.27
pip~=23.2.1
attrs~=23.2.0
Jinja2~=3.1.3
setuptools~=65.5.0
uvicorn~=0.25.0